+++ src/main/java/ru/stepup/payments/Point.java
@@ -1,8 +1,8 @@
 package ru.stepup.payments;
 
 public class Point {
-    int x,y;
-    public Point(int x, int y) {
+    double x,y;
+    public Point(double x, double y) {
         this.x = x;
         this.y = y;
     }
@@ -12,11 +12,11 @@ public class Point {
         return "{" + x + "," + y + "}";
     }
 
-    public void setX(int x) {
+    public void setX(double x) {
         this.x = x;
     }
 
-    public void setY(int y) {
+    public void setY(double y) {
         this.y = y;
     }
 }
